pyproject.toml
[project]
name = "srag-opendatasus-challenge"
version = "0.1.0"
authors = [
    { name = "Henrique Bezerra Lucas", email = "henriquelucas15@outlook.com.br" }
]
readme = "README.md"
requires-python = ">=3.11.9,<3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.5",
    "bleach==6.2.0",
    "certifi==2025.8.3",
    "cffi>=2.0.0b1",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dacite==1.9.2",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "dotenv==0.9.9",
    "executing==2.2.0",
    "fastjsonschema==2.21.2",
    "fonttools==4.59.1",
    "fqdn==1.5.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "htmlmin==0.1.12",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "imagehash==4.3.1",
    "ipykernel==6.30.1",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "json5==0.12.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.6",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.6",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.15",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-core==0.3.74",
    "langchain-openai==0.3.31",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.6.6",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.3",
    "langsmith==0.4.17",
    "lark==1.2.2",
    "llvmlite==0.44.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "multimethod==1.12",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "notebook==7.4.5",
    "notebook-shim==0.2.4",
    "numba==0.61.0",
    "numpy==2.1.3",
    "openai==1.101.0",
    "ordered-set==4.1.0",
    "orjson==3.11.2",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "patsy==1.0.1",
    "phik==0.12.5",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "psycopg2-binary==2.9.10",
    "pure-eval==0.2.3",
    "puremagic==1.30",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pylatex==1.4.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-json-logger==3.3.0",
    "pytz==2025.2",
    "pywavelets==1.9.0",
    "pywin32==311",
    "pywinpty==3.0.0",
    "pyyaml==6.0.2",
    "pyzmq==27.0.2",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rpds-py==0.27.0",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.43",
    "stack-data==0.6.3",
    "statsmodels==0.14.5",
    "tavily-python==0.7.11",
    "tenacity==9.1.2",
    "terminado==0.18.1",
    "tiktoken==0.11.0",
    "tinycss2==1.4.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typeguard==4.4.4",
    "types-python-dateutil==2.9.0.20250822",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "visions==0.8.1",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "widgetsnbextension==4.0.14",
    "wordcloud==1.9.4",
    "xxhash==3.5.0",
    "ydata-profiling==4.16.1",
    "zstandard==0.24.0",
]
